Stop exporting bash PS1.
While PS1 is a bash-only variable, a subshell like bash-acting-as-sh spawned with `sudo su` seems to use PS1 definition.
In this case, as the shell is acting as sh, it won't read .bashrc, and fail to find the `prompt` function with the error:
> `sh: prompt: command not found`
Note that bash subshells will read ~/.bashrc and so use this PS1 definition anyway. eg., bash shell spawned by `nix develop`